Move ChartData
This wasn't working with the initializer so instead I just manually
imported it into each route, which I think is a bit cleaner anyway.
