+++ tests/dummy/app/lib/chart-data.js
