+++ tests/dummy/app/routes/chart-bar.js
@@ -1,4 +1,8 @@
 import Ember from 'ember';
+import ChartData from 'dummy/lib/chart-data';
 
 export default Ember.Route.extend({
+  model() {
+    return ChartData.get('bar');
+  }
 });
