+++ tests/dummy/app/routes/index.js
@@ -3,6 +3,6 @@ import ChartData from 'dummy/lib/chart-data';
 
 export default Ember.Route.extend({
   model() {
-    return ChartData.get('donut');
+    return ChartData;
   }
 });
