+++ tests/dummy/app/templates/chart-horz-bar.hbs
@@ -5,15 +5,15 @@
 
 <div class='detail-chart'>
   <div class='icon-example example'>
-    {{simple-chart isIcon=true name='horz-bar' data=chartData.horz}}
+    {{simple-chart isIcon=true name='horz-bar' data=model}}
   </div>
 
   <div class='medium-example example'>
-    {{simple-chart name='horz-bar' data=chartData.horz}}
+    {{simple-chart name='horz-bar' data=model}}
   </div>
 
   <div class='large-example example'>
-    {{#simple-chart name='horz-bar' data=chartData.horz hover=(action (mut data)) as |chart|}}
+    {{#simple-chart name='horz-bar' data=model hover=(action (mut data)) as |chart|}}
       {{#if data}}
         {{#chart.tooltip title=(concat data.label ' (' data.data ')')}}
           {{chartData.ipsum}}
