+++ tests/integration/components/simple-chart-horz-bar-test.js
@@ -1,6 +1,6 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import ChartData from '../../../tests/chart-data';
+import ChartData from 'dummy/lib/chart-data';
 
 moduleForComponent('simple-chart-horz-bar', 'Integration | Component | simple chart horz bar', {
   integration: true
