+++ tests/integration/components/simple-chart-pie-test.js
@@ -1,6 +1,6 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import ChartData from '../../../tests/chart-data';
+import ChartData from 'dummy/lib/chart-data';
 
 moduleForComponent('simple-chart-pie', 'Integration | Component | simple chart pie', {
   integration: true
